requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet == 3.0.4
colorama==0.4.1
Django==2.2.2
django-braces==1.13.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-embed-video==1.2.0
django-memcache-status==2.1
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
pylint==2.3.1
python-memcached==1.59
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
